+++ src/ts/toolbar/Emoji.ts
@@ -63,6 +63,9 @@ data-value=":${key}: " data-key=":${key}:" class="vditor-emojis__icon" src="${em
                 event.preventDefault();
                 const value = element.getAttribute("data-value");
                 if (vditor.currentMode === "wysiwyg") {
+                    if (!vditor.wysiwyg.element.contains(getSelection().getRangeAt(0).startContainer)) {
+                        vditor.wysiwyg.element.focus();
+                    }
                     const range = getSelection().getRangeAt(0);
                     if (value.indexOf(":") > -1) {
                         insertHTML(vditor.lute.SpinVditorDOM(value), vditor);