+++ src/ts/toolbar/Headings.ts
@@ -3,6 +3,7 @@ import {insertText} from "../editor/insertText";
 import {getEventName, updateHotkeyTip} from "../util/compatibility";
 import {removeHeading, setHeading} from "../wysiwyg/setHeading";
 import {MenuItem} from "./MenuItem";
+import {afterRenderEvent} from "../wysiwyg/afterRenderEvent";
 
 export class Headings extends MenuItem {
     public element: HTMLElement;
@@ -13,12 +14,12 @@ export class Headings extends MenuItem {
 
         const headingsPanelElement = document.createElement("div");
         headingsPanelElement.className = "vditor-hint vditor-arrow";
-        headingsPanelElement.innerHTML = `<button data-value="# ">Heading 1 ${updateHotkeyTip("<⌘-⌥-1>")}</button>
-<button data-value="## ">Heading 2 ${updateHotkeyTip("<⌘-⌥-2>")}</button>
-<button data-value="### ">Heading 3 ${updateHotkeyTip("<⌘-⌥-3>")}</button>
-<button data-value="#### ">Heading 4 ${updateHotkeyTip("<⌘-⌥-4>")}</button>
-<button data-value="##### ">Heading 5 ${updateHotkeyTip("<⌘-⌥-5>")}</button>
-<button data-value="###### ">Heading 6 ${updateHotkeyTip("<⌘-⌥-6>")}</button>`;
+        headingsPanelElement.innerHTML = `<button data-tag="h1" data-value="# ">Heading 1 ${updateHotkeyTip("&lt;⌘-⌥-1>")}</button>
+<button data-tag="h2" data-value="## ">Heading 2 ${updateHotkeyTip("&lt;⌘-⌥-2>")}</button>
+<button data-tag="h3" data-value="### ">Heading 3 ${updateHotkeyTip("&lt;⌘-⌥-3>")}</button>
+<button data-tag="h4" data-value="#### ">Heading 4 ${updateHotkeyTip("&lt;⌘-⌥-4>")}</button>
+<button data-tag="h5" data-value="##### ">Heading 5 ${updateHotkeyTip("&lt;⌘-⌥-5>")}</button>
+<button data-tag="h6" data-value="###### ">Heading 6 ${updateHotkeyTip("&lt;⌘-⌥-6>")}</button>`;
 
         this.element.appendChild(headingsPanelElement);
 
@@ -30,6 +31,7 @@ export class Headings extends MenuItem {
             const actionBtn = this.element.children[0];
             if (vditor.currentMode === "wysiwyg" && actionBtn.classList.contains("vditor-menu--current")) {
                 removeHeading(vditor);
+                afterRenderEvent(vditor);
             } else {
                 if (headingsPanelElement.style.display === "block") {
                     headingsPanelElement.style.display = "none";
@@ -50,7 +52,8 @@ export class Headings extends MenuItem {
         for (let i = 0; i < 6; i++) {
             headingsPanelElement.children.item(i).addEventListener(getEventName(), (event: Event) => {
                 if (vditor.currentMode === "wysiwyg") {
-                    setHeading(vditor, (event.target as HTMLElement).tagName.toLowerCase());
+                    setHeading(vditor, (event.target as HTMLElement).getAttribute("data-tag"));
+                    afterRenderEvent(vditor);
                 } else {
                     insertText(vditor, (event.target as HTMLElement).getAttribute("data-value"), "",
                         false, true);