+++ CHANGELOG.md
@@ -45,10 +45,11 @@
 * [55](https://github.com/Vanessa219/vditor/issues/55) 链接引用不转换为内联链接 `改进功能`
 * [81](https://github.com/Vanessa219/vditor/issues/81) 链接和图片嵌套问题 `修复缺陷`
 
-* [148](https://github.com/Vanessa219/vditor/issues/148) checkbox with del key `修复缺陷`
+* [138](https://github.com/Vanessa219/vditor/issues/138) ctrl+b, ctrl+i, enter with no text `修复缺陷`
 
 ### v2.2.1 / 2020-02-18
 
+* [148](https://github.com/Vanessa219/vditor/issues/148) 任务列表退格删除问题 `修复缺陷`
 * [141](https://github.com/Vanessa219/vditor/issues/141) toolbar 添加箭头，默认表情修改 `改进功能`
 * [140](https://github.com/Vanessa219/vditor/issues/140) *** after shift+enter `修复缺陷`
 * [139](https://github.com/Vanessa219/vditor/issues/139) toggle after ===, enter `修复缺陷`